+++ example/e2e/testCases/HookInitialValue.spec.web.js
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @format
+ */
+/* eslint-env jest */
+
+describe('HookInitialValue', () => {
+  it('should have e2e tests in the future', async () => {
+    await expect(true).toEqual(true);
+  });
+});
