+++ package.json
@@ -64,7 +64,6 @@
     "@types/react-native": "^0.57.51",
     "@typescript-eslint/eslint-plugin": "^1.7.0",
     "@typescript-eslint/parser": "^1.7.0",
-    "babel-jest": "^24.7.1",
     "babel-loader": "^8.0.6",
     "detox": "^12.8.0",
     "eslint": "5.16.0",
@@ -73,6 +72,8 @@
     "html-webpack-plugin": "^3.2.0",
     "husky": "^2.3.0",
     "jest": "^24.8.0",
+    "jest-expo": "^36.0.1",
+    "jest-fetch-mock": "^3.0.1",
     "lint-staged": "^8.1.7",
     "metro-react-native-babel-preset": "^0.56.0",
     "prettier": "^1.17.0",
