+++ src/__tests__/fetch.spec.ts
@@ -8,17 +8,28 @@
  */
 
 import NetInfo from '../index';
-import {NativeModules} from 'react-native';
 import NativeInterface from '../internal/nativeInterface';
 import {NetInfoStateType, NetInfoCellularGeneration} from '../internal/types';
 import {DEVICE_CONNECTIVITY_EVENT} from '../internal/privateTypes';
+import nativeModule from '../internal/nativeModule';
+
+// Mock modules
+require('jest-fetch-mock').enableMocks();
+jest.mock('./../internal/nativeModule', () =>
+  require('./../../jest/nativeModule.mock'),
+);
 
 describe('@react-native-community/netinfo fetch', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
   describe('with cellular data types', () => {
     describe('4g cellular generation', () => {
       beforeEach(() => {
         jest.clearAllMocks();
-        NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue({
+        // @ts-ignore
+        nativeModule.getCurrentState.mockResolvedValue({
           type: NetInfoStateType.cellular,
           isConnected: true,
           isInternetReachable: true,
@@ -74,7 +85,8 @@ describe('@react-native-community/netinfo fetch', () => {
 
       dataProvider().forEach(testCase => {
         it(testCase.description, () => {
-          NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue(
+          // @ts-ignore
+          nativeModule.getCurrentState.mockResolvedValue(
             testCase.expectedConnectionInfo,
           );
 
@@ -92,7 +104,8 @@ describe('@react-native-community/netinfo fetch', () => {
     describe('3g cellular generation', () => {
       beforeEach(() => {
         jest.clearAllMocks();
-        NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue({
+        // @ts-ignore
+        nativeModule.getCurrentState.mockResolvedValue({
           type: NetInfoStateType.cellular,
           isConnected: true,
           isInternetReachable: true,
@@ -148,7 +161,8 @@ describe('@react-native-community/netinfo fetch', () => {
 
       dataProvider().forEach(testCase => {
         it(testCase.description, () => {
-          NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue(
+          // @ts-ignore
+          nativeModule.getCurrentState.mockResolvedValue(
             testCase.expectedConnectionInfo,
           );
 
@@ -166,7 +180,8 @@ describe('@react-native-community/netinfo fetch', () => {
     describe('2g cellular generation', () => {
       beforeEach(() => {
         jest.clearAllMocks();
-        NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue({
+        // @ts-ignore
+        nativeModule.getCurrentState.mockResolvedValue({
           type: NetInfoStateType.cellular,
           isConnected: true,
           isInternetReachable: true,
@@ -222,7 +237,8 @@ describe('@react-native-community/netinfo fetch', () => {
 
       dataProvider().forEach(testCase => {
         it(testCase.description, () => {
-          NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue(
+          // @ts-ignore
+          nativeModule.getCurrentState.mockResolvedValue(
             testCase.expectedConnectionInfo,
           );
 
@@ -240,7 +256,8 @@ describe('@react-native-community/netinfo fetch', () => {
     describe('null cellular generation', () => {
       beforeEach(() => {
         jest.clearAllMocks();
-        NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue({
+        // @ts-ignore
+        nativeModule.getCurrentState.mockResolvedValue({
           type: NetInfoStateType.cellular,
           isConnected: true,
           isInternetReachable: true,
@@ -296,7 +313,8 @@ describe('@react-native-community/netinfo fetch', () => {
 
       dataProvider().forEach(testCase => {
         it(testCase.description, () => {
-          NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue(
+          // @ts-ignore
+          nativeModule.getCurrentState.mockResolvedValue(
             testCase.expectedConnectionInfo,
           );
 
@@ -315,7 +333,8 @@ describe('@react-native-community/netinfo fetch', () => {
   describe('with unknown and no connection data types', () => {
     beforeEach(() => {
       jest.clearAllMocks();
-      NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue({
+      // @ts-ignore
+      nativeModule.getCurrentState.mockResolvedValue({
         type: NetInfoStateType.cellular,
         isConnected: true,
         isInternetReachable: true,
@@ -352,7 +371,8 @@ describe('@react-native-community/netinfo fetch', () => {
 
     dataProvider().forEach(testCase => {
       it(testCase.description, () => {
-        NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue(
+        // @ts-ignore
+        nativeModule.getCurrentState.mockResolvedValue(
           testCase.expectedConnectionInfo,
         );
 
@@ -370,7 +390,8 @@ describe('@react-native-community/netinfo fetch', () => {
   describe('rejected promises are handled', () => {
     beforeEach(() => {
       jest.clearAllMocks();
-      NativeModules.RNCNetInfo.getCurrentState.mockResolvedValue({
+      // @ts-ignore
+      nativeModule.getCurrentState.mockResolvedValue({
         type: NetInfoStateType.cellular,
         isConnected: true,
         isInternetReachable: true,
@@ -381,19 +402,17 @@ describe('@react-native-community/netinfo fetch', () => {
       });
     });
 
-    it('will properly reject a promise if the connection request cannot be resolved', () => {
+    it('will properly reject a promise if the connection request cannot be resolved', async () => {
       const rejectionMessage = 'nope, no connection info for you';
 
-      NativeModules.RNCNetInfo.getCurrentState.mockRejectedValue(
-        rejectionMessage,
-      );
+      // @ts-ignore
+      nativeModule.getCurrentState.mockRejectedValue(rejectionMessage);
 
-      NativeInterface.eventEmitter.emit(
-        DEVICE_CONNECTIVITY_EVENT,
-        rejectionMessage,
-      );
-
-      expect(NetInfo.fetch()).rejects.toMatch(rejectionMessage);
+      try {
+        await NetInfo.fetch();
+      } catch (error) {
+        expect(error.message).toMatch(rejectionMessage);
+      }
     });
   });
 });
